+++ src/rockverse/_assert/condition.py
@@ -1,5 +1,5 @@
 import numpy as np
-from rockverse._assert.utils import collective_raise
+from rockverse.errors import collective_raise
 
 def integer_or_float(varname, var):
     if np.dtype(type(var)).kind in 'fui': #splitting to handle >= op only if number
@@ -11,6 +11,12 @@ def non_negative_integer_or_float(varname, var):
         if var>=0:
             return
     collective_raise(ValueError(f"Expected non negative number for {varname}."))
+
+def positive_integer_or_float(varname, var):
+    if np.dtype(type(var)).kind in 'fui': #splitting to handle >= op only if number
+        if var>0:
+            return
+    collective_raise(ValueError(f"Expected positive number for {varname}."))
 
 def integer(varname, var):
     if np.dtype(type(var)).kind in 'ui':