+++ src/rockverse/dualenergyct/_cpu_functions.py
@@ -1,9 +1,9 @@
 import numpy as np
 from numba import njit
-from mpi4py import MPI
-comm = MPI.COMM_WORLD
-mpi_rank = comm.Get_rank()
-mpi_nprocs = comm.Get_size()
+from rockverse.config import config
+comm = config.mpi_comm
+mpi_rank = config.mpi_rank
+mpi_nprocs = config.mpi_nprocs
 
 from rockverse.dualenergyct._corefunctions import calcABn_cpu
 from rockverse.dualenergyct._corefunctions import calc_rho_Z_cpu