+++ src/rockverse/errors.py
@@ -1,3 +1,7 @@
+'''
+errors module cannot call rockverse config due to circular imports.
+Call directly MPI from mpi4py
+'''
 import sys
 from mpi4py import MPI
 comm = MPI.COMM_WORLD
@@ -7,4 +11,5 @@ mpi_nprocs = comm.Get_size()
 def collective_raise(e):
     if mpi_rank == 0:
         raise e
-    sys.exit(1)
+    else:
+        sys.exit(1)